requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
build==0.10.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
dnspython==2.3.0
docopt==0.6.2
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
langchain==0.0.177
llama-cpp-python==0.1.53
loguru==0.7.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pinecone-client==3.0.0rc2
pip-tools==6.13.0
pipreqs==0.4.13
playwright==1.33.0
pydantic==1.10.7
pyee==9.0.4
pyproject_hooks==1.0.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gsmmodem-new==0.13.0
PyYAML==6.0
redis==4.5.5
regex==2023.5.5
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
tenacity==8.2.2
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
wikipedia==1.4.0
yarg==0.1.9
yarl==1.9.2
